Refactor bootstrap process and update configuration files
