+++ .env.example
@@ -1,18 +1,43 @@
-# FastAPI Configuration
-FASTAPI_HOST=0.0.0.0
-FASTAPI_PORT=8000
-
 # Kafka Configuration
 KAFKA_BOOTSTRAP_SERVERS=kafka1:19092
 KAFKA_CRYPTO_TOPIC=crypto_prices
-KAFKA_TOPIC_CONFIG_FILE=kafka.yaml
-KAFKA_DEFAULT_PARTITIONS=1
+KAFKA_TOPIC_CONFIG_FILE=bootstrap/topics.yaml
+KAFKA_DEFAULT_PARTITIONS=3
 KAFKA_DEFAULT_REPLICATION=1
 KAFKA_PRODUCER_CLIENT_ID=crypto_producer
+KAFKA_PRODUCER_VALUE_SERIALIZER=json
+ZOOKEEPER_CONNECT=zoo1:2181
 
-# ksqlDB Configuration
-KSQLDB_URL=http://ksqldb-server:8088
+# Kafka Broker Configuration
+KAFKA_BROKER_ID=1
+KAFKA_INTERNAL_BROKER_LISTENER=INTERNAL
+KAFKA_INTERNAL_BROKER_PORT=19092
+KAFKA_EXTERNAL_BROKER_PORT=9092
+KAFKA_DOCKER_BROKER_PORT=29092
 
-# General Configuration
+# Schema Registry Configuration
+SCHEMA_REGISTRY_URL=http://kafka-schema-registry:8081
+SCHEMA_REGISTRY_HOST=kafka-schema-registry
+SCHEMA_REGISTRY_PORT=8081
+
+# Kafka Connect Configuration
+KAFKA_CONNECT_URL=http://kafka-connect:8083
+KAFKA_CONNECT_GROUP_ID=compose-connect-group
+KAFKA_CONNECT_CONFIG_STORAGE_TOPIC=docker-connect-configs
+KAFKA_CONNECT_OFFSET_STORAGE_TOPIC=docker-connect-offsets
+KAFKA_CONNECT_STATUS_STORAGE_TOPIC=docker-connect-status
+
+# KSQL Configuration
+KSQL_URL=http://ksqldb-server:8088
+KSQL_SERVICE_ID=ksqldb-server_
+
+# Kafka REST Proxy Configuration
+KAFKA_REST_PROXY_URL=http://kafka-rest-proxy:8082
+
+# Application Configuration
+FASTAPI_HOST=0.0.0.0
+FASTAPI_PORT=8000
 PYTHONUNBUFFERED=1
+
+# Docker Configuration
 DOCKER_HOST_IP=127.0.0.1
